transcribe_reader/pyproject.toml
[project]
name = "transcribe-reader"
version = "0.1.0"
description = "Sync VTT transcription files from S3 to GitLab"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "sqlalchemy>=2.0",
    "pyodbc",
    "python-dotenv",
    "pydantic",
    "boto3",
    "dependency-injector",
    "python-gitlab>=4.0",
]

[project.scripts]
transcribe-reader = "transcribe_reader.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/transcribe_reader"]
